legal-doc-platform/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
celery==5.3.4
redis==5.0.1
django-ratelimit==4.1.0
django-audit-log==2.0.0
cryptography==41.0.8
django-extensions==3.2.3
django-filter==23.5
django-oauth-toolkit==1.7.1
django-allauth==0.57.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-debug-toolbar==4.2.0
pytest-django==4.7.0
factory-boy==3.3.0
black==23.12.0
flake8==6.1.0
isort==5.13.2
